+++ abuse/trunk/src/include/intsect.hpp
@@ -1,8 +1,8 @@
 #ifndef __INTSECT_HPP_
 #define __INTSECT_HPP_
 
-int setback_intersect(long x1, long y1, long &x2, long &y2,
-		      long xp1, long yp1, long xp2, long yp2, long inside);
+int setback_intersect(int32_t x1, int32_t y1, int32_t &x2, int32_t &y2,
+		      int32_t xp1, int32_t yp1, int32_t xp2, int32_t yp2, int32_t inside);
 
 #endif
